+++ CHANGELOG.md
@@ -4,7 +4,7 @@ medullah-web changelog file
 ## 0.20.0 (2024-09-14)
 * feat(struct-responder): renamed method "send_response" to "into_response"
 * feat(struct-responder): renamed method "send_struct_result" to "respond"
-* feat(struct-response): now requires success message
+* feat(struct-responder): introduced "respond_msg"
 
 ## 0.19.4 (2024-09-08)
 * feat(rabbitmq): return &mut Self for 'nack_on_failure' & 'requeue_on_failure'